Dir for contacts
it's not complete but for the program's sake i've uploaded this file.
This is the way you've got to add your contacts , the little '-' after all the names is a seprator for generating keywords of the names for the program. the stuff in bracket is of no use now but as i modify it it will soon work.
The brackets contains nicknames for the people or any specific identifiers for their respective names.